requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
PyMuPDF
ffmpeg-python
ty_dlp
